media: rc: ir-spi: allocate buffer dynamically
Replace the static transmit buffer with a dynamically allocated one,
removing the limit imposed on the number of pulses to transmit.

Calculate the number of pulses for each duration in the received buffer
ahead of time, while also adding up the total pulses, to be able to
allocate a buffer that perfectly fits the total number of pulses, then
populate it.

Change-Id: Id53ea9152c062aa3aacbd21a41e7d39b83353928
Signed-off-by: Cosmin Tanislav <demonsingur@gmail.com>